+++ DB/src/main/java/ies/jandula/universidad/repository/AsignaturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ies.jandula.universidad.models.Asignatura;
 
-public interface AsignaturaRepository extends JpaRepository<Asignatura, Long>{
+public interface AsignaturaRepository extends JpaRepository<Asignatura, Integer>{
 
 }
